# ReplitExport-leasemasterpulszip/ReplitExport-bam254thzip/Pharma-POS-Manager/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PharmaPOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory with stock levels
      parameters:
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/{productId}/adjust:
    post:
      operationId: adjustStock
      tags: [inventory]
      summary: Adjust stock level for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustment"
      responses:
        "200":
          description: Adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"

  # Customers
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}/transactions:
    get:
      operationId: getCustomerTransactions
      tags: [customers]
      summary: Get transaction history for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  # Transactions / Checkout
...
  /dashboard/revenue:
    get:
      operationId: getRevenueTrend
      tags: [dashboard]
      summary: Get revenue trend data for charts
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month]
            default: week
      responses:
        "200":
          description: Revenue trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenuePeriod"
...
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Get top selling products
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month]
            default: week
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
...
components:
  schemas:
...
    InventoryItem:
      type: object
      required: [productId, productName, sku, stockQty, lowStockThreshold, lastUpdated]
      properties:
        productId:
          type: integer
        productName:
          type: string
        sku:
          type: string
        category:
          type: string
        stockQty:
          type: integer
        lowStockThreshold:
          type: integer
        unit:
          type: string
        lastUpdated:
          type: string
        status:
          type: string
          enum: [ok, low, out]
    StockAdjustment:
      type: object
      required: [type, quantity]
      properties:
        type:
          type: string
          enum: [add, subtract, set]
        quantity:
          type: integer
        reason:
          type: string
    Customer:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        loyaltyPoints:
          type: integer
        totalSpend:
          type: number
        visitCount:
          type: integer
        createdAt:
          type: string
        lastVisit:
          type: ["string", "null"]
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
    Transaction:
      type: object
      required: [id, totalAmount, status, paymentMethod, createdAt, items]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        discountAmount:
          type: number
        taxAmount:
          type: number
        paidAmount:
          type: number
        changeAmount:
          type: number
        status:
          type: string
          enum: [pending, completed, refunded, cancelled]
        paymentMethod:
          type: string
          enum: [cash, card, mobile_money, other]
        referenceCode:
          type: ["string", "null"]
        validationCode:
          type: ["string", "null"]
        isValidated:
          type: boolean
        receiptPrinted:
          type: boolean
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransactionItem"
    TransactionItem:
      type: object
      required: [productId, productName, quantity, unitPrice, totalPrice]
      properties:
        productId:
          type: integer
        productName:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
...
    RevenuePeriod:
      type: object
      required: [label, revenue, transactions]
      properties:
        label:
          type: string
        revenue:
          type: number
        transactions:
          type: integer
...
    TopProduct:
      type: object
      required: [productId, productName, totalSold, totalRevenue]
      properties:
        productId:
          type: integer
        productName:
          type: string
        category:
          type: string
        totalSold:
          type: integer
        totalRevenue:
          type: number
...