# ReplitExport-leasemasterpulszip/ReplitExport-bam254thzip/Pharma-POS-Manager/lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaPOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  loyaltyPoints?: number;
  totalSpend?: number;
  visitCount?: number;
  createdAt: string;
  /** @nullable */
  lastVisit?: string | null;
}