# ReplitExport-leasemasterpulszip/ReplitExport-bam254thzip/Pharma-POS-Manager/lib/api-zod/src/generated/types/stockAdjustment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaPOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StockAdjustmentType } from './stockAdjustmentType';

export interface StockAdjustment {
  type: StockAdjustmentType;
  quantity: number;
  reason?: string;
}