# ReplitExport-leasemasterpulszip/ReplitExport-bam254thzip/Pharma-POS-Manager/lib/api-zod/src/generated/types/topProduct.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaPOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TopProduct {
  productId: number;
  productName: string;
  category?: string;
  totalSold: number;
  totalRevenue: number;
}