# ReplitExport-leasemasterpulszip/ReplitExport-bam254thzip/Pharma-POS-Manager/lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmaPOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionItem } from './transactionItem';
import type { TransactionPaymentMethod } from './transactionPaymentMethod';
import type { TransactionStatus } from './transactionStatus';

export interface Transaction {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  totalAmount: number;
  discountAmount?: number;
  taxAmount?: number;
  paidAmount?: number;
  changeAmount?: number;
  status: TransactionStatus;
  paymentMethod: TransactionPaymentMethod;
  /** @nullable */
  referenceCode?: string | null;
  /** @nullable */
  validationCode?: string | null;
  isValidated?: boolean;
  receiptPrinted?: boolean;
  createdAt: string;
  items: TransactionItem[];
}